// src/model/store/country-code.ts
export enum CountryCode {
  // North America
  CA = 'ca',
  MX = 'mx',
  US = 'us',
  // South America
  BR = 'br',
  AR = 'ar',
  // Europe
  BE = 'be',
  CZ = 'cz',
  DK = 'dk',
  DE = 'de',
  EE = 'ee',
  ES = 'es',
  FI = 'fi',
  FR = 'fr',
  GR = 'gr',
  HU = 'hu',
  IE = 'ie',
  IT = 'it',
  LV = 'lv',
  LT = 'lt',
  LU = 'lu',
  NL = 'nl',
  NO = 'no',
  AT = 'at',
  PL = 'pl',
  PT = 'pt',
  RO = 'ro',
  RU = 'ru',
  CH = 'ch',
  SI = 'si',
  SK = 'sk',
  SE = 'se',
  GB = 'gb',
  // Asia - Pacific
  AU = 'au',
  IN = 'in',
  MY = 'my',
  NZ = 'nz',
  SG = 'sg',
  KR = 'kr',
  CN = 'cn',
  JP = 'jp',
  HK = 'hk',
  TW = 'tw',
  // Orient
  ZA = 'za',
  TR = 'tr',
  AE = 'ae'
}